// frontend/src/backend.d.ts
export interface UserPreferences {
    customLogoPath?: string;
    customAppName?: string;
    dayMode: boolean;
}
export interface TransformationOutput {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export type Time = bigint;
export interface PriceChange {
    timeframe: string;
    percentage: number;
}
export interface ChartDataResponse {
    source: ApiSource;
    data: Array<ChartDataPoint>;
    timestamp: Time;
    priceChange: PriceChange;
}
export interface ApiResponse {
    status: ApiStatus;
    message: string;
    timestamp: Time;
}
export interface MetricDataPoint {
    value: number;
    timestamp: Time;
}
export interface OrderBookEntry {
    source: ApiSource;
    timestamp: Time;
    price: number;
    amount: number;
}
export interface ChartDataPoint {
    low: number;
    high: number;
    close: number;
    open: number;
    volume: number;
    timestamp: Time;
}
// ...
export interface ConversionRate {
    date: string;
    rate: number;
}
export enum ApiSource {
    internetComputerDashboard = "internetComputerDashboard",
    icpTokensNet = "icpTokensNet",
    coinGecko = "coinGecko",
    coinMarketCap = "coinMarketCap"
}
export enum ApiStatus {
    pending = "pending",
    error = "error",
    success = "success"
}
export enum ChartTimeframe {
    fourHours = "fourHours",
    oneHour = "oneHour",
    sevenDays = "sevenDays",
    oneYear = "oneYear",
    eightHours = "eightHours",
    fifteenMinutes = "fifteenMinutes",
    oneMonth = "oneMonth",
    oneDay = "oneDay"
}
export enum MetricType {
    canisterStorage = "canisterStorage",
    cycleBurnRate = "cycleBurnRate",
    icpCyclesConversion = "icpCyclesConversion",
    transactionsPerSecond = "transactionsPerSecond",
    canisterCount = "canisterCount",
    subnetCount = "subnetCount",
    nodeProviderCount = "nodeProviderCount",
    nodeMachineCount = "nodeMachineCount"
}